requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Django==2.1.7
django-background-tasks==1.2.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-compat==1.0.15
django-workers==0.1.2
idna==2.8
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.6.8
pyTelegramBotAPI==3.6.6
pytz==2018.9
requests==2.21.0
service-identity==18.1.0
six==1.12.0
urllib3==1.24.1
wincertstore==0.2
